+++ docs/index.service.worker.js
@@ -4,7 +4,7 @@
 // Incrementing CACHE_VERSION will kick off the install event and force
 // previously cached resources to be updated from the network.
 /** @type {string} */
-const CACHE_VERSION = '1789803891|2318201';
+const CACHE_VERSION = '1789805808|1872773';
 /** @type {string} */
 const CACHE_PREFIX = 'Taterland-sw-cache-';
 const CACHE_NAME = CACHE_PREFIX + CACHE_VERSION;